Make base class for run job management commands
