+++ data_aggregator/management/commands/run_assignment_jobs.py
@@ -1,59 +1,25 @@
-import logging
-import traceback
-from django.core.management.base import BaseCommand
 from django.db import transaction
 from data_aggregator.dao import CanvasDAO
-from data_aggregator.models import Job, Assignment
+from data_aggregator.models import Assignment
+from data_aggregator.management.commands._base import RunJobCommand
 
 
-class Command(BaseCommand):
+class Command(RunJobCommand):
+
+    job_type = "assignment"
 
     help = ("Loads the assignment data for a batch of jobs. Designed to "
             "be run as a cron that is constantly checking for new jobs.")
 
-    def add_arguments(self, parser):
-        parser.add_argument("--job_batch_size",
-                            type=int,
-                            help=("Number of jobs to process"),
-                            default=10,
-                            required=False)
-
     @transaction.atomic
     def run_job(self, job):
-        try:
-            # delete existing assignment data in case of a job restart
-            old_assignments = Assignment.objects.filter(job=job)
-            old_assignments.delete()
-            # load assignment data
-            job.mark_start()
-            canvas_course_id = job.context["canvas_course_id"]
-            assignments = (
-                CanvasDAO().get_assignments(canvas_course_id))
-            for assign in assignments:
-                assign.job = job
-            Assignment.objects.bulk_create(assignments)
-            job.mark_end()
-        except Exception:
-            # save error message if one occurs
-            tb = traceback.format_exc()
-            job.message = tb
-            job.save()
-            self.logger.error(tb)
-
-    def handle(self, *args, **options):
-        """
-        Queries the Job model to check for unstarted jobs (jobs
-        where pid=None and start=None). For a batch of unstarted jobs
-        (default 10), queries the Canvas API for assignment data and stores
-        the returned data as Assignment model instances.
-        """
-        self.logger = logging.getLogger(__name__)
-        job_batch_size = options["job_batch_size"]
-        jobs = Job.objects.start_batch_of_assignment_jobs(
-            batchsize=job_batch_size
-        )
-        if jobs:
-            for job in jobs:
-                self.run_job(job)
-        else:
-            self.logger.info("No active assignment jobs.")
+        # delete existing assignment data in case of a job restart
+        old_assignments = Assignment.objects.filter(job=job)
+        old_assignments.delete()
+        # load assignment data
+        canvas_course_id = job.context["canvas_course_id"]
+        assignments = (
+            CanvasDAO().get_assignments(canvas_course_id))
+        for assign in assignments:
+            assign.job = job
+        Assignment.objects.bulk_create(assignments)
